MAINT: Move ma benchmarks to new file
