+++ benchmarks/benchmarks/bench_core.py
@@ -96,21 +96,6 @@ class Temporaries(Benchmark):
         (self.alarge + self.blarge) - 2
 
 
-class MA(Benchmark):
-    def setup(self):
-        self.l100 = range(100)
-        self.t100 = ([True] * 100)
-
-    def time_masked_array(self):
-        np.ma.masked_array()
-
-    def time_masked_array_l100(self):
-        np.ma.masked_array(self.l100)
-
-    def time_masked_array_l100_t100(self):
-        np.ma.masked_array(self.l100, self.t100)
-
-
 class CorrConv(Benchmark):
     params = [[50, 1000, 1e5],
               [10, 100, 1000, 1e4],
